Redo game loop with timeout and recursion.
I altered the game loop to use setTimeout to rerender the game once a second.
I used recursion with this to simulate a while loop that has a base case of
the game over conditions being met which is checked with a function. I did
this as a while loop would block the single thread of the runtime.